+++ docs/ChangeLog.md
@@ -150,6 +150,8 @@
 - contrib/fzf-git: fix unsupported command modes (reported by dgudim) `#D2083` ba2b8865
 - contrib/bash-preexec: support the latest version of `bash-preexec` (reported by mcarans) `#D2128` 50af4d9c
 - contrib/config/execmark: output error status through `ble/canvas/trace` `#D2136` 64cdcd01
+- contrib/airline: remove dummy faces (reported by alexalekseyenko) `#D2201` xxxxxxxx
+- contrib/airline: update themes `#D2201` xxxxxxxx
 
 ## Documentation